# openapi.yaml
openapi: 3.0.1
info:
  title: 항해플러스 E-Commerce API
  description: 항해플러스 이커머스 서비스 API 문서
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/coupons/issue:
    post:
      tags:
      - 쿠폰 API
      summary: 쿠폰 발급
      description: 사용자에게 쿠폰을 발급한다.
      operationId: issueCoupon
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCouponRequest"
        required: true
      responses:
        "201":
          description: 쿠폰 발급 성공
          content:
            application/json:
              examples:
                성공:
                  description: 성공
                  value:
                    code: 201
                    message: 쿠폰이 성공적으로 발급되었습니다
                    data:
                      userCouponId: 1
                      userId: 1
                      couponId: 1
                      couponName: 20% 할인 쿠폰
                      discountRate: 20
                      status: AVAILABLE
                      issuedAt: 2025-10-30T10:00:00
                      expiresAt: 2025-11-30T10:00:00
        "409":
          description: 발급 불가
          content:
            application/json:
              examples:
                이미 발급받음:
                  description: 이미 발급받음
                  value:
                    code: 409
                    message: 이미 발급받은 쿠폰입니다.
                    data: null
                쿠폰 소진:
                  description: 쿠폰 소진
                  value:
                    code: 409
                    message: 쿠폰이 모두 소진되었습니다.
                    data: null
        "404":
          description: 쿠폰 없음
          content:
            application/json:
              examples:
                쿠폰 없음:
                  description: 쿠폰 없음
                  value:
                    code: 404
                    message: 쿠폰을 찾을 수 없습니다.
                    data: null
...
  /api/coupons:
    get:
      tags:
      - 쿠폰 API
      summary: 쿠폰 조회
      description: 본인에게 발급된 쿠폰을 확인할 수 있다.
      operationId: getCoupons
      parameters:
      - name: userId
        in: query
        description: 사용자 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: status
        in: query
        description: 쿠폰 상태
        required: false
        schema:
          type: string
          enum:
          - AVAILABLE
          - USED
          - EXPIRED
        example: AVAILABLE
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCouponListResponse"
...
components:
  schemas:
...
    IssueCouponRequest:
      required:
      - couponId
      - userId
      type: object
      properties:
        userId:
          type: integer
          description: 사용자 ID
          format: int64
          example: 1
        couponId:
          type: integer
          description: 쿠폰 ID
          format: int64
          example: 1
      description: 쿠폰 발급 요청
...
    ApiResponseCouponListResponse:
      type: object
      properties:
        code:
          type: integer
          description: HTTP 상태 코드
          format: int32
          example: 200
        message:
          type: string
          description: 응답 메시지
          example: 요청이 정상적으로 처리되었습니다.
        data:
          $ref: "#/components/schemas/CouponListResponse"
      description: 공통 API 응답
    CouponListResponse:
      type: object
      properties:
        coupons:
          type: array
          description: 사용자 쿠폰 목록
          items:
            $ref: "#/components/schemas/UserCouponDto"
      description: 쿠폰 목록 조회 응답
    UserCouponDto:
      type: object
      properties:
        userCouponId:
          type: integer
          description: 사용자 쿠폰 ID
          format: int64
          example: 1
        couponId:
          type: integer
          description: 쿠폰 ID
          format: int64
          example: 1
        couponName:
          type: string
          description: 쿠폰명
          example: 20% 할인 쿠폰
        discountRate:
          type: integer
          description: 할인율 (%)
          format: int32
          example: 20
        status:
          type: string
          description: "쿠폰 상태 (AVAILABLE, USED, EXPIRED)"
          example: AVAILABLE
          enum:
          - AVAILABLE
          - USED
          - EXPIRED
        issuedAt:
          type: string
          description: 발급 시간
          format: date-time
        usedAt:
          type: string
          description: 사용 시간 (사용한 경우만)
          format: date-time
        expiresAt:
          type: string
          description: 만료 시간
          format: date-time
      description: 사용자 쿠폰 정보
...